Track Codex terminal event ownership across suspended writes
